Run `bin/rails app:update` and merge with HCB's configuration
The general philosophy is to make the config look like as if we had
started HCB with rails new on 8.0 and then applied the changes we did
over HCB's lifetime to get it to the current state. This meant deleting
default values (especially commented out ones) if they were removed in
newer versions of rails and copying over comment blocks to match rails 8.0
generators.